grow floating window when the text is too long
When the input is centered then the growth happens so that the window
is kept centered (grows on the right and shift the float to the left).

When the input is created relative to the cursor position (i.e. when
it's triggered by LSP rename symbol), the window grows only towards the
right side, keeping the column relative position below the renamed
symbol.